users can delete their own comments but not the ones from other users

+++ src/components/CommentCard.tsx
@@ -1,9 +1,9 @@
 import { Button, Card } from "react-bootstrap";
 import { CommentType } from "../types/customTypes";
-import { collection, deleteDoc, doc, getDocs } from "firebase/firestore";
+import { deleteDoc, doc } from "firebase/firestore";
 import { db } from "../config/firebaseConfig";
 import { useParams } from "react-router";
-import { useContext, useEffect, useState } from "react";
+import { useContext } from "react";
 import { AuthContext } from "../context/AuthContext";
 
 type CommentCardProps = {
@@ -12,57 +12,46 @@ type CommentCardProps = {
 
 function CommentCard({ comment }: CommentCardProps) {
   const { countryName } = useParams<string>();
-  const { user } = useContext(AuthContext);
+
+  const {user} = useContext(AuthContext)
 
   const dateFormat = (seconds: number) => {
     const formattedDate = new Date(seconds * 1000).toLocaleString();
     return formattedDate;
   };
 
+  console.log(user?.email);
+  console.log(comment.user.email);
+
   const handleCommentDelete = async (commentId: string) => {
     if (!countryName) {
       throw new Error("countryName is undefined!");
     }
 
-    try {
-      const commentDocRef = doc(
-        db,
-        "comments",
-        countryName,
-        "messages",
-        commentId
-      );
-      await deleteDoc(commentDocRef);
-      console.log("Message deleted with ID:", commentId);
-    } catch (error) {
-      console.error("Error deleting message:", error);
-    }
-  };
-
-  const [username, setUsername] = useState<string | null>(null);
-
-  const getUsername = async () => {
-    const querySnapshot = await getDocs(collection(db, "users"));
-    querySnapshot.forEach((doc) => {
-      if (user && user.id === doc.id) {
-        const displayName = doc.data().displayName;
-        setUsername(displayName);
-
-        if (!displayName) {
-          setUsername("friend");
-        }
+    if (user?.email === comment.user.email) {
+      try {
+        const commentDocRef = doc(
+          db,
+          "comments",
+          countryName,
+          "messages",
+          commentId
+        );
+        await deleteDoc(commentDocRef);
+        console.log("Message deleted with ID:", commentId);
+      } catch (error) {
+        console.error("Error deleting message:", error);
       }
-    });
-  };
+    };
+    }
 
-  useEffect(() => {
-    getUsername();
-  }, [user]); // Run the effect when the user changes
+
+
 
   return (
     <Card style={{ width: "18rem" }} key={comment.id}>
       <Card.Body>
-        <Card.Title>{username}</Card.Title>
+        <Card.Title>{comment.user.email}</Card.Title>
         <Card.Subtitle className="mb-2 text-muted">
           {dateFormat(comment.date.seconds)}
         </Card.Subtitle>
@@ -70,7 +59,7 @@ function CommentCard({ comment }: CommentCardProps) {
       </Card.Body>
       <div>
         <Button
-          className="comment-delete-button"
+          className={user?.email === comment.user.email ? "comment-delete" : "comment-delete-disabled"}
           onClick={() => handleCommentDelete(comment.id)}
         >
           Delete
